Initial commit of user-friendly FormantNet executable scripts
This is the first commit of Python scripts that can be used on the command line
to train FormantNet models and/or use them on test files:

FNet_train.py: Train a FormantNet model on provided wavefiles
FNet_track.py: Run a trained model on provided wavefiles

or, alternatively,

FormantNet.py: Do both (train a model and use it on test files)

Command-line syntax and options can be displayed with the -h option, e.g.:
  python3 FormantNet.py -h

The other python files (FN_*.py) are support files that are also needed to run
the scripts. An optional configuration file can be provided; FNet_config0.txt is
an example configuration file that lists all of the available options and their
default values. Example filelists for TIMIT are also provided. Notebook versions
of the Python files are stored in notebooks/ and are not needed to run the
scripts.

TO DO: Add a README.md file to describe these scripts in detail and all of the
available options.