More detailed time reporting
Modified FN_data.getdata() so that it will print a progress report
after about every 10% of the full data list (including the amount
of time taken so far).

Modified FN_model.track_files() to print a report of the total time
taken to run the test files (and average seconds per file).